refactor(transformer/object-rest-spread): move plugin-relates files to `object_rest_spread` mod (#5320)
Both `object_spread` and `object_rest` belong to `object_rest_object`, so moving them to the plugin mod makes the file structure clearer
