+++ crates/oxc_transformer/src/es2018/object_rest_spread/object_rest.rs
@@ -24,8 +24,9 @@
 //! * Babel plugin implementation: <https://github.com/babel/babel/tree/main/packages/babel-plugin-transform-object-rest-spread>
 //! * Object rest/spread TC39 proposal: <https://github.com/tc39/proposal-object-rest-spread>
 
-use super::object_rest_spread::ObjectRestSpreadOptions;
 use crate::context::Ctx;
+
+use super::ObjectRestSpreadOptions;
 
 pub struct ObjectRest<'a> {
     _ctx: Ctx<'a>,
